lowlevel

+++ 0x00-hello_world/6-size.c
@@ -1,16 +1,21 @@
 #include <stdio.h>
+
 /**
- * main - Entry Point
+ * main - print the size of various types
  *
- * Return: Always 0 ( Success )
+ * Description: prints the size of various types on the computer it is compiled
+ * and run on
+ *
+ * Return: Allways 0 (Success)
  */
 
 int main(void)
 {
-	printf("Size of a char: %ld byte(s)\n", sizeof(char));
-	printf("Size of an int: %ld byte(s)\n", sizeof(int));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(long));
-	printf("Size of a long long int: %ld byte(s)\n", sizeof(double));
-	printf("Size of a float: %ld byte(s)\n", sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int) sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of a long int: %d byte(s)\n", (int) sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n", (int) sizeof(float));
+
 	return (0);
 }
